+++ frontend/src/pages/Messages/Messages.tsx
@@ -1,21 +1,15 @@
 import React from "react";
 import { MessageDto } from "../../types/message/message.dto";
 import { Message } from "./Message";
+import { Socket } from "socket.io-client";
 
-export const Messages: React.FC<{wsChannel: WebSocket | null}> = ({wsChannel}) => {
+export const Messages: React.FC<{socket: Socket | null}> = ({ socket  }) => {
 
 	const [messages, setMessages] = React.useState<MessageDto[]>([]);
 
 	React.useEffect(() => {
-		const messageHandler = (e: MessageEvent) => {
-			const newMessages = JSON.parse(e.data);
-			setMessages((prevMessages) => [...prevMessages, ...newMessages])
-		};
-		wsChannel?.addEventListener('message', messageHandler)
 
-		return () => wsChannel?.removeEventListener('message', messageHandler)
-
-	},[wsChannel])
+	},[])
 	return <div style={{height: '400px', overflowY: 'auto'}}>
 		{messages.map((m, index) => <Message key={index} message={m}/>)}