+++ frontend/src/pages/Messages/MessagesPage.tsx
@@ -4,6 +4,10 @@ import { useDispatch } from "react-redux";
 import { socket } from "../../api/message.api";
 import { ConversationDto } from "../../types/message/conversation.dto";
 import { Button } from "@mui/material";
+import { Messages } from "./Messages";
+import { AddMessageForm } from "./AddMessage";
+import { Conversations } from "./Conversations";
+import { MessageDto } from "../../types/message/message.dto";
 
 interface MessagesPageInterface {
 	isAuth: boolean;
@@ -12,19 +16,26 @@ interface MessagesPageInterface {
 export const MessagesPage: React.FC<MessagesPageInterface> = ({ isAuth }) => {
 
 	const [conversations, setConversations] = React.useState<ConversationDto[]>([]);
+	const [messages, setMessages] = React.useState<MessageDto[]>([]);
 
 	React.useEffect(() => {
 		if (!isAuth) {
 			return navigate("/login");
 		}
 		 socket.emit('joinConversation', { friendId: 3})
-		 socket.emit('sendMessage', {  message: "Hello", userId: 1, conversationId: 3 })
+		// socket.emit('sendMessage', {  message: "Hello", userId: 1, conversationId: 3 })
 		// socket.emit('createConversation', { friendId: 3})
-		socket.emit('leaveConversation')
+		// socket.emit('leaveConversation')
 		socket.on("conversations", (data) => {
 			console.log(data);
 			setConversations([...conversations, data]);
 		});
+
+		socket.on("messages", (data) => {
+			console.log(data);
+			setMessages([...messages, data]);
+		});
+
 
 		return () => {
 			setConversations([]);
@@ -42,7 +53,9 @@ export const MessagesPage: React.FC<MessagesPageInterface> = ({ isAuth }) => {
 
 	return (
 		<div>
-			<Button>Create conversation</Button>
+			<Conversations conversations={conversations}/>
+			<Messages socket={socket}/>
+			<AddMessageForm socket={socket}/>
 		</div>
 	);
 };